Split /files endpoint to individual controllers
* Splits `FileController` to single-action controllers in `Files\`
* Updates test filenames for /files to match the new controller names
* Drops `PathDeletionFailed` exception from `FileManager::delete()`
* Adds tests for...
  * catching permission errors, ensuring we get a 403
  * deleting a directory (previously only file deletion was tested)
  * calling `mix()` (not `asset()`) via `smart_asset()` in local env
