--- app/FileManager/PathDeletionFailed.php
@@ -1,9 +1,0 @@
-<?php
-
-namespace Servidor\FileManager;
-
-use Exception;
-
-class PathDeletionFailed extends Exception
-{
-}
