+++ app/Http/Controllers/Files/Controller.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace Servidor\Http\Controllers\Files;
+
+use Servidor\FileManager\FileManager;
+use Servidor\Http\Controllers\Controller as BaseController;
+
+class Controller extends BaseController
+{
+    protected FileManager $fm;
+
+    public function __construct(FileManager $manager)
+    {
+        $this->fm = $manager;
+    }
+}
