+++ tests/Unit/Http/FallbackRouteTest.php
@@ -14,10 +14,25 @@ class FallbackRouteTest extends TestCase
     /** @test */
     public function frontend_routes_use_vue_layout(): void
     {
+        $this->app['env'] = 'testing';
+
         $response = $this->get('/');
 
         $response->assertViewIs('servidor');
         $response->assertSee('router-view id="app"');
+        $response->assertSee('/js/app.js');
+    }
+
+    /** @test */
+    public function app_layout_uses_mix_in_local_environment(): void
+    {
+        $this->app['env'] = 'local';
+
+        $response = $this->get('/');
+
+        $response->assertViewIs('servidor');
+        $response->assertSee('router-view id="app"');
+        $response->assertSee('/js/app.js');
     }
 
     /** @test */
